+++ src/callFactory.js
@@ -2,11 +2,12 @@ import {validateCreator, validateDefinition} from './utils/validate';
 import {createLogger} from './utils/logging';
 import createActions from './createActions';
 
-export default function createCall(name, {
+export default function callFactory(name, {
     namespace='global',
     defaultActions=['loading', 'error', 'success'],
     actions=createActions(`${namespace}:${name}`, defaultActions),
     logger=createLogger(`${namespace}:${name}`),
+    ...initialDefinition
 }={}) {
     const error = validateCreator({name, actions, logger}, logger);
     if (error) {
@@ -18,7 +19,7 @@ export default function createCall(name, {
         name,
         actions,
 
-        define: (definition) => {
+        create: (definition) => {
             const createDefinition = typeof definition === 'function' ? definition : () => definition;
             const call = Object.assign(createDefinition({name, actions, logger}), {
                 name,
@@ -29,7 +30,9 @@ export default function createCall(name, {
                 throw new Error('Invalid call');
             }
             call.dataSource = {
+                // set the default actions
                 ...(call.actions || {}),
+                // set dataSource from the definition passed om - potentially overriding the actions
                 ...(call.dataSource || {})
             };
             return call;